[refactor](be) Refine expression zonemap helper boundaries
### What problem does this PR solve?

Issue Number: None

Related PR: #63389

Problem Summary: This change keeps expression zonemap pruning helper logic focused and reusable. It centralizes single-slot zonemap candidate detection, scopes HybridSet literal materialization to VDirectInPredicate, removes redundant defensive handling after can_evaluate_zonemap_filter has already validated the expression shape, asserts the segment zonemap invariant at the ColumnReader accessor boundary, and drops an unrelated runtime filter unit-test change from the PR diff.

### Release note

None

### Check List (For Author)

- Test: Unit Test; Code style
    - `./build-support/clang-format.sh be/src/exprs/expr_zonemap_filter.cpp be/src/exprs/expr_zonemap_filter.h be/src/exprs/vdirect_in_predicate.h be/src/exprs/vexpr.cpp be/src/exprs/vexpr.h be/src/exprs/vin_predicate.cpp be/src/storage/segment/column_reader.cpp be/src/storage/segment/segment.cpp be/src/storage/segment/segment_iterator.cpp be/test/exec/runtime_filter/runtime_filter_consumer_helper_test.cpp`
    - `git diff --check`
    - `git diff --cached --check`
    - `./run-be-ut.sh --run --filter='ExprZonemapFilterTest.*:ParquetExprTest.test_expr_zonemap_*'`
- Behavior changed: No
- Does this need documentation: No
