+++ be/src/exprs/vexpr.h
@@ -642,9 +642,6 @@ Status create_texpr_literal_node(const void* data, TExprNode* node, int precisio
 TExprNode create_texpr_node_from(const void* data, const PrimitiveType& type, int precision = 0,
                                  int scale = 0);
 
-TExprNode create_texpr_node_from_hybrid_set_value(const void* data, const PrimitiveType& type,
-                                                  int precision = 0, int scale = 0);
-
 TExprNode create_texpr_node_from(const Field& field, const PrimitiveType& type, int precision,
                                  int scale);
